% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_length.R
\name{boot_length}
\alias{boot_length}
\title{resample length data w/replacement}
\usage{
boot_length(lfreq_un)
}
\arguments{
\item{lfreq_un}{expanded length frequency data}
}
\description{
resample length data w/replacement
}
