% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess_age.R
\name{ess_age}
\alias{ess_age}
\title{calculate effective sample size for age comps}
\usage{
ess_age(sim_data, og_data, strata)
}
\arguments{
\item{sim_data}{list of abundance by age data}

\item{og_data}{original abundance by age data (single list)}

\item{strata}{switch to perform analysis at regional or strata level (default = NULL)}
}
\description{
calculate effective sample size for age comps
}
