% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess_size.R
\name{ess_size}
\alias{ess_size}
\title{calculate effective sample size for length comps}
\usage{
ess_size(sim_data, og_data, strata)
}
\arguments{
\item{sim_data}{list of abundance by length data}

\item{og_data}{original abundance by length data (single list)}

\item{strata}{switch to perform analysis at regional or strata level (default = NULL)}
}
\description{
calculate effective sample size for length comps
}
