% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_reduce.R
\name{l_reduce}
\alias{l_reduce}
\title{reduce total number of lengths}
\usage{
l_reduce(data, samples, grp)
}
\arguments{
\item{data}{lfreq dataframe}

\item{samples}{number of lengths to take}

\item{write_sample}{whether to output intermediate step}

\item{region}{region being sampled}

\item{save}{name to save}
}
\description{
reduce total number of lengths
}
