% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpop.R
\name{lpop}
\alias{lpop}
\title{population at length}
\usage{
lpop(lcomp, cpue, lngs)
}
\arguments{
\item{lcomp}{length comp}

\item{cpue}{cpue data}

\item{lngs}{complete lengths by year}
}
\description{
population at length
}
