% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_samples.R
\name{reduce_samples}
\alias{reduce_samples}
\title{reduce total number of lengths or ages}
\usage{
reduce_samples(
  data,
  samples,
  grp = c("year", "species_code", "stratum", "hauljoin")
)
}
\arguments{
\item{data}{lfreq or afreq dataframe}

\item{samples}{number of lengths to take}

\item{grp}{grouping variable(s)}
}
\description{
reduce total number of lengths or ages
}
