% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swo.R
\name{swo}
\alias{swo}
\title{primary survey workload optimization function}
\usage{
swo(
  lfreq_data,
  specimen_data,
  cpue_data,
  strata_data,
  yrs = NULL,
  strata = FALSE,
  boot_hauls = FALSE,
  boot_lengths = FALSE,
  boot_ages = FALSE,
  length_samples = NULL,
  sample_sex = NULL
)
}
\arguments{
\item{lfreq_data}{length frequency data}

\item{specimen_data}{age-length specimen data}

\item{cpue_data}{abundance by length data}

\item{strata_data}{strata and associated area}

\item{yrs}{age filter (returns years >=)}

\item{strata}{switch for analyzing by strata (default = NULL) - not currently implemented}

\item{boot_hauls}{switch for resampling hauls (default = NULL)}

\item{boot_lengths}{switch for resampling lengths (default = NULL)}

\item{boot_ages}{switch for resampling ages (default = NULL)}

\item{length_samples}{change sample sizes (default = NULL)}

\item{sample_sex}{change sample sizes (default = NULL)}
}
\value{

}
\description{
primary survey workload optimization function
}
\examples{
swo(lfreq, specimen, cpue, strata_data, yrs = 2015, boot_hauls = TRUE,
    boot_lengths = TRUE, length_samples = 100)
}
