% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swo_ai_rebs.R
\name{swo_ai_rebs}
\alias{swo_ai_rebs}
\title{primary survey workload optimization function (customized for ai rebs complex)}
\usage{
swo_ai_rebs(
  lfreq_data,
  specimen_data,
  cpue_data,
  strata_data,
  yrs,
  strata,
  boot_hauls,
  boot_lengths,
  boot_ages,
  reduce_lengths,
  length_samples,
  sex_samples
)
}
\arguments{
\item{lfreq_data}{length frequency data}

\item{specimen_data}{age-length specimen data}

\item{cpue_data}{abundance by length data}

\item{strata_data}{strata and associated area}

\item{yrs}{age filter returns years >= (default = NULL)}

\item{strata}{switch for analyzing by strata (default = FALSE) - not currently implemented}

\item{boot_hauls}{switch for resampling hauls (default = FALSE)}

\item{boot_lengths}{switch for resampling lengths (default = FALSE)}

\item{boot_ages}{switch for resampling ages (default = FALSE)}

\item{reduce_lengths}{reduce the total number of lengths used in the analysis (default = NULL)}

\item{length_samples}{change sample sizes (default = NULL)}

\item{sex_samples}{change sample sizes (default = NULL)}
}
\description{
primary survey workload optimization function (customized for ai rebs complex)
}
\examples{
swo(lfreq, specimen, cpue, strata_data, yrs = 2015, boot_hauls = TRUE,
    boot_lengths = TRUE, length_samples = 100)
}
