% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swo_tables.R
\name{swo_tables}
\alias{swo_tables}
\title{two basic swo tables}
\usage{
swo_tables(lfreq, spec, region)
}
\arguments{
\item{lfreq}{length frequency data}

\item{spec}{species data}

\item{region}{region of interes}
}
\description{
two basic swo tables
}
\examples{
\dontrun{ 
swo_tables(lfreq, spec, 'goa')
}
}
