% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample}
\alias{sample}
\title{resample length or sex data}
\usage{
sample(lfreq_un, samples, type = "length")
}
\arguments{
\item{lfreq_un}{expanded length frequency data}

\item{samples}{number of samples to take (default = NULL)}

\item{type}{length or sex samples (default = 'length')}
}
\description{
resample length or sex data
}
