% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swo.R
\name{swo}
\alias{swo}
\title{primary survey workload optimization function}
\usage{
swo(
  lfreq_data,
  specimen_data,
  cpue_data,
  strata_data,
  yrs,
  strata,
  boot_hauls,
  boot_lengths,
  boot_ages,
  length_samples,
  sex_samples,
  write_sample,
  save,
  region
)
}
\arguments{
\item{lfreq_data}{length frequency data}

\item{specimen_data}{age-length specimen data}

\item{cpue_data}{abundance by length data}

\item{strata_data}{strata and associated area}

\item{yrs}{age filter returns years >= (default = NULL)}

\item{strata}{switch for analyzing by strata (default = FALSE) - not currently implemented}

\item{boot_hauls}{switch for resampling hauls (default = FALSE)}

\item{boot_lengths}{switch for resampling lengths (default = FALSE)}

\item{boot_ages}{switch for resampling ages (default = FALSE)}

\item{length_samples}{change sample sizes (default = NULL)}

\item{sex_samples}{change sample sizes (default = NULL)}

\item{write_sample}{switch to save the "new_unsexed" data (default = FALSE)}

\item{save}{name to save a file}

\item{region}{region will create a folder and place results in said folder}
}
\value{

}
\description{
primary survey workload optimization function
}
\examples{
swo(lfreq, specimen, cpue, strata_data, yrs = 2015, boot_hauls = TRUE,
    boot_lengths = TRUE, length_samples = 100)
}
