% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swo_sim.R
\name{swo_sim}
\alias{swo_sim}
\title{replicate swo function and generate input sample sizes}
\usage{
swo_sim(
  iters = 1,
  lfreq_data,
  specimen_data,
  cpue_data,
  strata_data,
  yrs = NULL,
  strata = FALSE,
  boot_hauls = FALSE,
  boot_lengths = FALSE,
  boot_ages = FALSE,
  length_samples = NULL,
  age_samples = NULL,
  sexlen_samples = NULL,
  save = NULL,
  write_interm = FALSE,
  region = NULL
)
}
\arguments{
\item{iters}{number of iterations (500 recommended)}

\item{lfreq_data}{input dataframe}

\item{specimen_data}{input dataframe}

\item{cpue_data}{input dataframe}

\item{strata_data}{input dataframe}

\item{yrs}{any year filter >= (default = NULL)}

\item{strata}{switch for regional or by strata (default = FALSE)}

\item{boot_hauls}{resample hauls w/replacement (default = FALSE)}

\item{boot_lengths}{resample lengths w/replacement (default = FALSE)}

\item{boot_ages}{resample ages w/replacement (default = FALSE)}

\item{length_samples}{sample size by length (default = NULL)}

\item{age_samples}{proportion of sample size by age (default = NULL)}

\item{sexlen_samples}{sample size by sex-length (default = NULL)}

\item{save}{name to save a file}

\item{write_interm}{save the intermediate age/length comps, the new "unsexed/unsampled" samples, and original length/age pop'n ests}

\item{region}{region will create a folder and place results in said folder}
}
\value{
effective sample sizes and imput sample sizes
}
\description{
replicate swo function and generate input sample sizes
}
\examples{
\dontrun{
swo_sim(iters=500, lfreq_data, specimen_data, cpue_data, strata_data, yrs = 2015,
                   strata = FALSE, boot_hauls = TRUE, boot_lengths = TRUE, boot_ages = FALSE,
                  length_samples = NULL, age_samples = NULL, sexlen_samples = NULL, save = "test", 
                   write_interm = FALSE, region = "goa")
}
}
